Fixed spike bug in charmm2.

+++ uniform/kernels.h
@@ -50,8 +50,8 @@ public:
   real_t (*globMultipole)[MTERM];
   real_t (*globLocal)[LTERM];
   int (*Leafs)[2];
-  float (*sendJbodies)[4];
-  float (*recvJbodies)[4];
+  double (*sendJbodies)[4];
+  double (*recvJbodies)[4];
   float (*sendMultipole)[MTERM];
   float (*recvMultipole)[MTERM];
   int (*sendLeafs)[2];
